Questions, Survey Creation, Answer Survey Pages are working with separated user DB's

+++ src/main/java/com/example/questions/service/QuestionService.java
@@ -17,14 +17,15 @@ public class QuestionService {
     private QuestionRepository repository;
 
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_SURVEYCREATOR')")
-    public boolean createQuestion(QuestionModel question) {
+    public boolean createQuestion(QuestionModel question, String userName) {
+       question.setUserName(userName);
         repository.save(question);
         return true;
     }
 
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_SURVEYCREATOR')")
-    public void deleteAllQuestions(){
-        repository.deleteAll();
+    public void deleteAllQuestions(String userName){
+        repository.findAll().stream().filter(x -> x.getUserName().equals(userName)).forEach(x -> repository.deleteById(x.getId()));
     }
 
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_SURVEYCREATOR')")
@@ -33,19 +34,31 @@ public class QuestionService {
     }
 
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_SURVEYCREATOR')")
-    public boolean deleteById(String id){
-        repository.deleteById(id);
-        return true;
+    public boolean deleteById(String id, String userName){
+        Optional<QuestionModel> q = repository.findById(id);
+        if(userName.equals(q.get().getUserName())){
+            repository.deleteById(id);
+            return true;
+        }else {
+            return false;
+        }
     }
 
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_SURVEYCREATOR')")
-    public Optional<QuestionModel> getQuestionById(String id){
-        return repository.findById(id);
+    public Optional<QuestionModel> getQuestionById(String id, String userName){
+        QuestionModel q = repository.findById(id).get();
+        if(userName.equals(q.getUserName())){
+            return repository.findById(id);
+        }
+        return null;
     }
 
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_SURVEYCREATOR')")
-    public void updateById(QuestionModel question, String id) {
-        repository.deleteById(id);
-        repository.save(question);
+    public void updateById(QuestionModel question, String id, String userName) {
+        QuestionModel q = repository.findById(id).get();
+        if(q.getUserName().equals(userName)){
+            repository.deleteById(id);
+            repository.save(question);
+        }
     }
 }
